+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Versions follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (`<ma
 ### Added
 * Add parsing pipelines for Flysight V1 track data files
 * #8 Add parsing pipelines for Flysight V2 track & sensor data files
-* #11 Add trinning for Flysight V1 & V2 data files
+* #11 Add trimming for Flysight V1 & V2 data files
+* #12 Add simplified CSV log export for Flysight V2 data files
 
 ## [v0.1.0]
 Initial release - yay!